Move FielsSelect to new Form folder and refactor
